+++ RuneScape/Utils.cs
@@ -128,7 +128,7 @@ namespace Supay.Bot {
       }
     }
 
-    public static int SoulWarsZealToExp(string skill, int startExp, int targetExp, bool bonus) {
+    public static int SoulWarsZealToExp(string skill, long startExp, long targetExp, bool bonus) {
       int zeal = 0;
       while (startExp < targetExp) {
         int expPerZeal = Utils.SoulWarsExpPerZeal(skill, startExp.ToLevel());
@@ -168,7 +168,7 @@ namespace Supay.Bot {
       return (int)Math.Ceiling((double)((level + 25) * (level - 24)) / 606.0) * modifier;
     }
 
-    public static int PestControlPointsToExp(string skill, int startExp, int targetExp, int bonus) {
+    public static int PestControlPointsToExp(string skill, long startExp, long targetExp, int bonus) {
       int points = 0;
       while (startExp < targetExp) {
         int expPerPoint = Utils.PestControlExpPerPoint(skill, startExp.ToLevel());
@@ -190,7 +190,7 @@ namespace Supay.Bot {
       return points;
     }
 
-    public static int LampsToExp(int startExp, int targetExp) {
+    public static int LampsToExp(long startExp, long targetExp) {
       int lamps = 0;
       while (startExp < targetExp) {
         startExp += 10 * Math.Min(startExp.ToLevel(), 99);
@@ -199,7 +199,7 @@ namespace Supay.Bot {
       return lamps;
     }
 
-    public static int BooksToExp(int startExp, int targetExp) {
+    public static int BooksToExp(long startExp, long targetExp) {
       int books = 0;
       while (startExp < targetExp) {
         startExp += 15 * Math.Min(startExp.ToLevel(), 99);
